rename connection workload
