+++ packages/backend/src/index.ts
@@ -12,7 +12,7 @@ import { shutdownPosthog } from "./posthog.js";
 import { prisma } from "./prisma.js";
 import { PromClient } from './promClient.js';
 import { redis } from "./redis.js";
-import { createConnectionWorkload } from "./connectionWorkload.js";
+import { createConnectionSyncWorkload } from "./connectionSyncWorkload.js";
 import { cleanupOrphanedRepoResources, createRepoIndexWorkload } from "./repoIndexWorkload.js";
 import { Api } from "./api.js";
 import { createAccountPermissionSyncWorkload } from "./ee/accountPermissionSyncWorkload.js";
@@ -50,7 +50,7 @@ logger.info('Worker started.');
 
 const jobManager = new BullMQJobManager(redis);
 
-const connectionWorkload = createConnectionWorkload({
+const connectionSyncWorkload = createConnectionSyncWorkload({
     db: prisma,
     jobManager,
     settings,
@@ -77,7 +77,7 @@ const auditLogPruneWorkload = createAuditLogPruneWorkload({
     retentionDays: env.SOURCEBOT_EE_AUDIT_RETENTION_DAYS,
 });
 
-jobManager.register(connectionWorkload);
+jobManager.register(connectionSyncWorkload);
 jobManager.register(repoIndexWorkload);
 jobManager.register(accountPermissionSyncWorkload);
 jobManager.register(repoPermissionSyncWorkload);
